Switches bulma external url
## Description
Worked with nuxt for the first time today and found this slight issue. After selecting `bulma` as my css framework of choice, it was weird as to why the external link provided lead to bootstrap-vue.

I made this PR to switch the link to the appropriate channel.

PS - the doesnt seem to be a `https://go.nuxtjs.dev/bulma` link so i just used the main bulma website instead. Feel free to change the link if you see fit 🙂

Cheers