fix(eslint): revert adding eslint-plugin-vue to devDependency
This reverts commit 7cef1461933464294857ebad31ea0574d2f23e29.

The eslint-plugin-vue isn't configured to be used by the .eslintrc.js,
so it's installed redundantly as long as only `prettier` rule is used.

If officially installed via `npx @vue/cli add @vue/cli-plugin-eslint`,
it adds ['plugin:vue/essential','eslint:recommended','@vue/prettier']
to the extends array; or it can be used with `plugin:vue/recommended`,
but it needs further analysis to avoid some other potential conflicts.